Got animation working and updated documentation
