+++ iOS/Inventory/Constants.swift
@@ -9,7 +9,6 @@
 import UIKit
 
 struct Constants {
-
     struct Colors {
         static let mainColor = UIColor(red: 0/255, green: 122/255, blue: 255/255, alpha: 1)
         static let lightHighlightColor = UIColor(red: 210/255, green: 224/255, blue: 253/255, alpha: 1)
