--- iOS/Inventory/Extensions.swift
@@ -1,16 +1,0 @@
-//
-//  Extensions.swift
-//  Inventory
-//
-//  Created by Shunsuke Kondo on 2023/01/23.
-//  Copyright © 2023 Ditto. All rights reserved.
-//
-
-import Foundation
-
-extension Array {
-
-    var indexes: [Int] {
-        return (Array<Int>)(indices)
-    }
-}
