+++ iOS/Inventory/Strings.swift
@@ -1,0 +1,20 @@
+//
+//  Strings.swift
+//  Inventory
+//
+//  Created by Erik Everson on 9/6/24.
+//  Copyright © 2024 Ditto. All rights reserved.
+//
+
+import Foundation
+
+// MARK: - Static keys
+
+let collectionNameKey = "inventories"
+let _idKey = "_id"
+let counterKey = "counter"
+let itemIdKey = "itemId"
+let imageNameKey = "imageName"
+let titleKey = "title"
+let priceKey = "price"
+let detailKey = "detail"
